refactor: move ReviewTable in table package
