+++ src/test/kotlin/se/uulm/snowballr/backend/repository/association/ReviewTableRepoTest.kt
@@ -13,8 +13,8 @@ import se.uulm.snowballr.backend.repository.RepositoryHelper.insertReviewAndGetI
 import se.uulm.snowballr.backend.repository.RepositoryTest
 import se.uulm.snowballr.backend.table.PaperTable
 import se.uulm.snowballr.backend.table.ProjectTable
+import se.uulm.snowballr.backend.table.ReviewTable
 import se.uulm.snowballr.backend.table.association.ProjectPaperTable
-import se.uulm.snowballr.backend.table.association.ReviewTable
 import snowballr.ReviewOuterClass
 import java.util.UUID
